Wrap SQL statement in a real class
And move it to another separate header
